backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-auth==2.23.4
nltk==3.8.1
numpy==1.26.1
networkx==3.2.1
python-dotenv==1.0.0